Prevent exceptions when saving associated models
The problem happens when persisting a relation to the model.
If the associated model is having a duplicate exception, the columns
against which the error occurred will not exist in the current model.

This causes problems even if the model didn't use 
`rescue_from_duplicate`.

Ensuring to return an empty array will let the error bubble.