Use Docker to setup the databases
Docker wasn't available when the gem was initially created. Complicated manual
setup was necessary to make sure all the tests ran properly.

Because of this we had elected to only run an easy subset of the tests (Sqlite3).
With the help of Docker we can now run all the tests all the time